# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Education CRM Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/teachers": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Get list of teachers",
                "description": "Returns list of teachers",
                "operationId": "2979badd3972683932b0d6d4d69900fa",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "skill_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "room_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "course_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+123456789"
                                                },
                                                "salary": {
                                                    "type": "integer",
                                                    "example": 50000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Create a new teacher",
                "operationId": "5e0f38e3163529ab40f4263056fce8ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "skill_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+123456789"
                                    },
                                    "salary": {
                                        "type": "integer",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Teacher created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "skill_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "room_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "course_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+123456789"
                                                },
                                                "salary": {
                                                    "type": "integer",
                                                    "example": 50000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/teachers/{teacherId}/add-course": {
            "post": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Add a course to a teacher",
                "description": "Adds a course to a specific teacher",
                "operationId": "02f5f31147ff469d111e9838e9914583",
                "parameters": [
                    {
                        "name": "teacherId",
                        "in": "path",
                        "description": "ID of the teacher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course added to teacher successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course added to teacher successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Course added to teacher successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course added to teacher successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid data provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Teacher or Course not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Teacher or Course not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/teachers/{id}": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Get a teacher by ID",
                "operationId": "c92f313e1b0f25be66d4c0a912e166c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the teacher to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+123456789"
                                                },
                                                "salary": {
                                                    "type": "integer",
                                                    "example": 50000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Update a teacher",
                "operationId": "b7aacebae3ed09ba7234de4f99bb422c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of teacher to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "skill_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+123456789"
                                    },
                                    "salary": {
                                        "type": "integer",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Teacher updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "skill_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "room_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "course_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+123456789"
                                                },
                                                "salary": {
                                                    "type": "integer",
                                                    "example": 50000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Delete a teacher",
                "operationId": "439c713bfcad2e10aa5495c70e691fe6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of teacher to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Teacher deleted successfully"
                    },
                    "404": {
                        "description": "Teacher not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}